+++ cmd/deploy.go
@@ -20,7 +20,8 @@ import (
 
 const (
 	VALUES_FLAG                   = "values"
-	CHART_NAME                    = "groundcover"
+	CHART_NAME                    = "groundcover/groundcover"
+	HELM_REPO_NAME                = "groundcover"
 	DEFAULT_GROUNDCOVER_RELEASE   = "groundcover"
 	DEFAULT_GROUNDCOVER_NAMESPACE = "groundcover"
 	COMMIT_HASH_KEY_NAME_FLAG     = "git-commit-hash-key-name"
@@ -275,9 +276,13 @@ func getClusterName(kubeClient *k8s.Client) (string, error) {
 
 func getLatestChart(helmClient *helm.Client, sentryHelmContext *sentry_utils.HelmContext) (*helm.Chart, error) {
 	var err error
-	var chart *helm.Chart
 
-	if chart, err = helmClient.GetLatestChart(CHART_NAME, HELM_REPO_URL); err != nil {
+	if err = helmClient.AddRepo(HELM_REPO_NAME, HELM_REPO_URL); err != nil {
+		return nil, err
+	}
+
+	var chart *helm.Chart
+	if chart, err = helmClient.GetLatestChart(CHART_NAME); err != nil {
 		return nil, err
 	}