+++ cmd/status.go
@@ -81,7 +81,7 @@ var StatusCmd = &cobra.Command{
 		sentry_utils.SetTagOnCurrentScope(sentry_utils.CHART_VERSION_TAG, sentryHelmContext.ChartVersion)
 
 		var chart *helm.Chart
-		if chart, err = helmClient.GetLatestChart(CHART_NAME, HELM_REPO_URL); err != nil {
+		if chart, err = getLatestChart(helmClient, sentryHelmContext); err != nil {
 			return err
 		}